Refactor the model generator, move the prompt and the schema repository out of the model generator
